+++ inlang/packages/paraglide/paraglide-js/src/services/logger/index.ts
@@ -22,7 +22,7 @@ export class Logger {
 
 	public log(message: string): Logger {
 		if (this.options.silent) return this;
-		const prefix = this.options.prefix ? colors.bold("[paraglide] ") : "";
+		const prefix = this.options.prefix ? colors.bold("[paraglide-js] ") : "";
 		consola.log(prefix + message);
 		return this;
 	}
@@ -30,7 +30,7 @@ export class Logger {
 	public info(message: string): Logger {
 		if (this.options.silent) return this;
 		const prefix = this.options.prefix
-			? colors.bold(colors.blue("[paraglide] "))
+			? colors.bold(colors.blue("[paraglide-js] "))
 			: "";
 		consola.info(prefix + message);
 		return this;
@@ -39,7 +39,7 @@ export class Logger {
 	public success(message: string): Logger {
 		if (this.options.silent) return this;
 		const prefix = this.options.prefix
-			? colors.bold(colors.green("[paraglide] "))
+			? colors.bold(colors.green("[paraglide-js] "))
 			: "";
 		consola.success(prefix + message);
 		return this;