+++ inlang/packages/paraglide/paraglide-js/CHANGELOG.md
@@ -4,6 +4,8 @@
 
 - rename `defineGetLocale()` and `defineSetLocale()` to `overwriteGetLocale()` and `overwriteSetLocale()` https://github.com/opral/inlang-paraglide-js/issues/382
 
+- enables `import { m } from "./paraglide/messages.js"` for auto imports https://github.com/opral/inlang-paraglide-js/issues/345
+
 ## 2.0.0-beta.18
 
 Added URLPatterns as a replacement for the beta 17 pathnames API.